+++ src/pages/Tags.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import styled from 'styled-components';
-import { ColumnFilterItem, FilterDropdownProps } from 'antd/lib/table/interface';
-import { Table, Tag as AntTag } from 'antd';
+import { FilterDropdownProps, TablePaginationConfig } from 'antd/lib/table/interface';
+import { SearchOutlined } from '@ant-design/icons';
+import { Table, Tag as AntTag, Space, Button } from 'antd';
 
 import { thunks } from '../../store';
 import { RootState, AppDispatch } from '../../store/types';
@@ -11,6 +12,7 @@ import { Tag, TagType } from '../../types/gelbooruTypes';
 import { capitalize, getTagColor } from '../../util/utils';
 import { unwrapResult } from '@reduxjs/toolkit';
 import TagSearchFilter from '../components/tags/TagSearchFilter';
+import TypeSearchFilter from '../components/tags/TypeSearchFilter';
 
 const { Column } = Table;
 
@@ -31,10 +33,11 @@ const Tags: React.FunctionComponent<Props> = (props: Props) => {
 
 	const [tagsCount, setTagsCount] = useState(0);
 	const [pattern, setPattern] = useState('');
-	const tagsPerPage = 19;
+	const [selectedTypes, setSelectedTypes] = useState<TagType[]>(['artist', 'character', 'copyright', 'metadata', 'tag']);
+	const [page, setPage] = useState<number>(1);
+	const tagsPerPage = 16;
 
 	useEffect(() => {
-		// dispatch(loadAllTagsFromDb());
 		dispatch(thunks.tags.getCount())
 			.then(unwrapResult)
 			.then((result) => setTagsCount(result));
@@ -42,17 +45,17 @@ const Tags: React.FunctionComponent<Props> = (props: Props) => {
 	}, []);
 
 	useEffect(() => {
-		dispatch(thunks.tags.getCount(pattern))
+		setPage(1);
+		dispatch(thunks.tags.getCount({ pattern, types: selectedTypes }))
 			.then(unwrapResult)
 			.then((result) => setTagsCount(result));
-		dispatch(thunks.tags.loadAllWithLimitAndOffset({ limit: tagsPerPage, offset: 0, pattern }));
-	}, [pattern]);
+		dispatch(thunks.tags.loadAllWithLimitAndOffset({ limit: tagsPerPage, offset: 0, pattern, types: selectedTypes }));
+	}, [pattern, selectedTypes]);
 
-	const handleChangePage = (page: number): void => {
-		dispatch(thunks.tags.loadAllWithLimitAndOffset({ limit: tagsPerPage, offset: (page - 1) * tagsPerPage, pattern }));
+	const handleChangePage = (p: number): void => {
+		setPage(p);
+		dispatch(thunks.tags.loadAllWithLimitAndOffset({ limit: tagsPerPage, offset: (p - 1) * tagsPerPage, pattern, types: selectedTypes }));
 	};
-
-	// const getFilteredTags = (): void => {};
 
 	const handleOnlineSearch = (tag: Tag): void => {
 		dispatch(thunks.tags.searchTagOnline(tag));
@@ -62,141 +65,76 @@ const Tags: React.FunctionComponent<Props> = (props: Props) => {
 		dispatch(thunks.tags.searchTagOffline(tag));
 	};
 
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
-	const renderActions = (_: unknown, record: Tag): JSX.Element => {
-		return (
-			<div>
-				<a
-					onClick={(): void => {
-						handleOnlineSearch(record);
-					}}
-				>
-					Online Search
-				</a>
-				<span> </span>
-				<a
-					onClick={(): void => {
-						handleOfflineSearch(record);
-					}}
-					style={{ float: 'right' }}
-				>
-					Offline Search
-				</a>
-			</div>
-		);
+	const tagSearchFilterOnSearch = (pattern: string): void => {
+		setPattern(pattern);
 	};
 
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
-	const generateFilterObject = (name: string, _value?: unknown): ColumnFilterItem => {
+	const renderTagSearchFilterDropdown = (dropdownProps: FilterDropdownProps): React.ReactNode => {
+		return <TagSearchFilter onSearch={tagSearchFilterOnSearch} confirm={dropdownProps.confirm} visible={dropdownProps.visible} />;
+	};
+
+	const typeSearchFilterOnSearch = (types: TagType[]): void => {
+		setSelectedTypes(types);
+	};
+
+	const renderTypeSearchFilterDropdown = (dropdownProps: FilterDropdownProps): React.ReactNode => {
+		return <TypeSearchFilter onSearch={typeSearchFilterOnSearch} confirm={dropdownProps.confirm} visible={true} />;
+	};
+
+	const getPaginationProps = (): TablePaginationConfig => {
 		return {
-			text: capitalize(name),
-			value: name,
+			pageSize: tagsPerPage,
+			total: tagsCount,
+			onChange: handleChangePage,
+			current: page,
+			showSizeChanger: false,
 		};
 	};
 
-	const getTypeFilters = (): ColumnFilterItem[] => {
-		return [
-			generateFilterObject('artist'),
-			generateFilterObject('character'),
-			generateFilterObject('copyright'),
-			generateFilterObject('metadata'),
-			generateFilterObject('tag'),
-		];
+	const renderActions = (_: unknown, record: Tag): JSX.Element => {
+		const handleOnline = (): void => {
+			handleOnlineSearch(record);
+		};
+
+		const handleOffline = (): void => {
+			handleOfflineSearch(record);
+		};
+
+		return (
+			<Space>
+				<Button type="link" onClick={handleOnline}>
+					Online Search
+				</Button>
+				<Button type="link" onClick={handleOffline}>
+					Offline Search
+				</Button>
+			</Space>
+		);
 	};
 
 	const renderTag = (_text: unknown, tag: Tag): JSX.Element => {
 		return <AntTag color={getTagColor(tag.type)}>{capitalize(tag.type)}</AntTag>;
 	};
 
-	const tagSearchFilterOnSearch = (pattern: string): void => {
-		setPattern(pattern);
-	};
-
-	const renderTagSearchFilterDropdown = (dropdownProps: FilterDropdownProps): React.ReactNode => {
-		return <TagSearchFilter onSearch={tagSearchFilterOnSearch} visible={dropdownProps.visible} />;
-	};
-
-	// const getAmbiguousFilters = (): ColumnFilterItem[] => {
-	// 	return [generateFilterObject('true'), generateFilterObject('false')];
-	// };
-
 	return (
 		<Container className={props.className}>
 			<Table
 				size="small"
 				dataSource={tags}
-				pagination={{
-					pageSize: tagsPerPage,
-					total: tagsCount,
-					onChange: handleChangePage,
-				}}
+				pagination={getPaginationProps()}
 				loading={tagsLoading}
 				rowKey="id"
+				bordered
 				rowClassName={(_record, index): string => (index % 2 === 0 ? 'table-row-light' : 'table-row-dark')}
 			>
-				<Column title="Id" dataIndex="id" width={100} /*sorter={(a: Tag, b: Tag): number => a.id - b.id} */ />
-				<Column
-					title="Tag"
-					dataIndex="tag"
-					ellipsis
-					filterDropdown={renderTagSearchFilterDropdown}
-					// sorter={(a: Tag, b: Tag): number => a.tag.localeCompare(b.tag)}
-				/>
-				<Column
-					title="Type"
-					dataIndex="type"
-					width={100}
-					onFilter={(value: TagType | string | number | boolean, record: Tag): boolean => record.type === value}
-					//filters={getTypeFilters()}
-					render={renderTag}
-				/>
-				<Column title="Count" dataIndex="count" width={100} /*sorter={(a: Tag, b: Tag): number => a.count - b.count}*/ />
-				<Column
-					title="Ambiguous"
-					dataIndex="ambiguous"
-					width={100}
-					render={(_text, record: Tag): string => (record.ambiguous == 0 ? 'false' : 'true')}
-					// onFilter={(value: string, record: Tag): boolean => {
-					// 	const numericValue = value === 'true' ? 1 : 0;
-					// 	console.log(`${record.ambiguous} === ${numericValue} | ${record.ambiguous == numericValue}`);
-					// 	return record.ambiguous === numericValue;
-					// }}
-					// filters={getAmbiguousFilters()}
-				/>
-				<Column
-					title="Favorites Count"
-					dataIndex="favoriteCount"
-					/*sorter={(a: Tag, b: Tag): number => {
-						if (a.favoriteCount !== undefined && b.favoriteCount !== undefined) {
-							return a.favoriteCount - b.favoriteCount;
-						} else {
-							return 0;
-						}
-					}}*/
-				/>
-				<Column
-					title="Blacklisted Count"
-					dataIndex="blacklistedCount"
-					/*sorter={(a: Tag, b: Tag): number => {
-						if (a.blacklistedCount !== undefined && b.blacklistedCount !== undefined) {
-							return a.blacklistedCount - b.blacklistedCount;
-						} else {
-							return 0;
-						}
-					}}*/
-				/>
-				<Column
-					title="Downloaded Count"
-					dataIndex="downloadedCount"
-					/*sorter={(a: Tag, b: Tag): number => {
-						if (a.downloadedCount !== undefined && b.downloadedCount !== undefined) {
-							return a.downloadedCount - b.downloadedCount;
-						} else {
-							return 0;
-						}
-					}}*/
-				/>
-				<Column title="Actions" dataIndex="" width={240} render={renderActions} />
+				<Column title="Id" dataIndex="id" width={100} />
+				<Column title="Tag" dataIndex="tag" ellipsis filterIcon={<SearchOutlined />} filterDropdown={renderTagSearchFilterDropdown} />
+				<Column title="Type" dataIndex="type" width={100} filterDropdown={renderTypeSearchFilterDropdown} render={renderTag} />
+				<Column title="Count" dataIndex="count" width={100} />
+				<Column title="Favorites Count" dataIndex="favoriteCount" />
+				<Column title="Blacklisted Count" dataIndex="blacklistedCount" />
+				<Column title="Downloaded Count" dataIndex="downloadedCount" />
+				<Column title="Actions" dataIndex="" width={260} render={renderActions} />
 			</Table>
 		</Container>
 	);