+++ changelog.md
@@ -1,9 +1,11 @@
 # Changelog
-## 0.1.3 - Settings are now in modal - DATE
+## 0.1.3 - New Tags page - DATE
 - Changed Settings to be contained in a Modal window
 - Added options to disable loading of dashboard statistics on App start
 - Add pagination to Tags page, disabled filters for now
 - Added Loader to FullSizeImage
+- Favorites directories can now be renamed
+- Tags page can now be searched by pattern and tag type with proper pagination
 - BUG FIX: Fixed most favorited tags and favorite post count not using data from the new tree structure.
 - BUG FIX: Fixed autoscrolling of thumbnails list in side panel